Combined offer a ride and request a ride to one java and one xml file.
